# account-service/doc/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "List of APIs for Micro Service",
        "title": "Micro Service API Document",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8082",
    "basePath": "/api/v1",
    "paths": {
...
        "/tenantRegister": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tenant"
                ],
                "summary": "Register tenant",
                "parameters": [
                    {
                        "description": "Tenant data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.TenantRegisterForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.TenantResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagdto.Error404"
                        }
                    }
                }
            }
        },
        "/tenants": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "List all existing tenants",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tenant"
                ],
                "summary": "List all existing tenants",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.TenantListResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagdto.Error404"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tenant"
                ],
                "summary": "Add tenant",
                "parameters": [
                    {
                        "description": "Tenant Data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.TenantForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.TenantResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagdto.Error404"
                        }
                    }
                }
            }
        },
        "/tenants/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tenant"
                ],
                "summary": "Get a tenant",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.TenantResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagdto.Error404"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tenant"
                ],
                "summary": "Update tenant",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Tenant ID",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.TenantForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.TenantResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagdto.Error404"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tenant"
                ],
                "summary": "Delete a tenant",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.TenantResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagdto.Error404"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tenant"
                ],
                "summary": "Patch tenant",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Tenant ID",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.TenantForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.TenantResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagdto.Error404"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.TenantDto": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "Tenant1"
                },
                "createdAt": {
                    "type": "string",
                    "example": "2021-02-02T02:52:24Z"
                },
                "domain": {
                    "type": "string",
                    "example": "EBOOK"
                },
                "email": {
                    "type": "string",
                    "example": "tenant@mail.com"
                },
                "id": {
                    "type": "integer",
                    "example": 123
                },
                "isActive": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string",
                    "example": "Tenant 1"
                },
                "updatedAt": {
                    "type": "string",
                    "example": "2021-02-02T02:52:24Z"
                }
            }
        },
        "model.TenantForm": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "Tenant1"
                },
                "domain": {
                    "type": "string",
                    "example": "EBOOK"
                },
                "email": {
                    "type": "string",
                    "example": "tenant@mail.com"
                },
                "name": {
                    "type": "string",
                    "example": "Tenant 1"
                }
            }
        },
        "model.TenantRegisterForm": {
            "type": "object",
            "properties": {
                "domain": {
                    "type": "string",
                    "example": "eBook"
                },
                "email": {
                    "type": "string",
                    "example": "tenant1@mail.com"
                },
                "firstName": {
                    "type": "string",
                    "example": "John"
                },
                "lastName": {
                    "type": "string",
                    "example": "Doe"
                },
                "name": {
                    "type": "string",
                    "example": "Tenant1"
                },
                "password": {
                    "type": "string",
                    "example": "Pass@1"
                }
            }
        },
...
        "swagdto.Error404": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "error": {
                    "$ref": "#/definitions/swagdto.ErrorNotFound"
                },
                "status": {
                    "type": "integer",
                    "example": 404
                }
            }
        },
...
        "swagdto.ErrorNotFound": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "NOT_FOUND"
                },
                "message": {
                    "type": "string",
                    "example": "Resource not found"
                }
            }
        },
...
        "swagger.TenantListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/swagger.TenantSampleListData"
                },
                "error": {
                    "type": "object"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
        "swagger.TenantResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/swagger.TenantSampleData"
                },
                "error": {
                    "type": "object"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
        "swagger.TenantSampleData": {
            "type": "object",
            "properties": {
                "tenant": {
                    "$ref": "#/definitions/model.TenantDto"
                }
            }
        },
        "swagger.TenantSampleListData": {
            "type": "object",
            "properties": {
                "tenants": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.TenantDto"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}